app/requirements.txt
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.4
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
Pillow==9.0.1
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
shellingham==1.4.0
sympy==1.12
tokenizers==0.15.0
toml==0.10.2
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typer==0.3.2
typing_extensions==4.9.0
urllib3==2.1.0
